pyproject.toml
[project]
name = "linear-enhancer"
version = "0.1.0"
description = "AI-powered issue creation from context across Slack, G Drive, and GitHub"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "openai-agents[litellm]>=0.6.4",
    "python-dotenv>=1.2.1",
    "slack-sdk>=3.39.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
